wm: Use size_t instead of int in tests.
Serves me right for TBR-ing the last change. :-P  Thanks!

Review URL: http://chromereview.prom.corp.google.com/1186082

git-svn-id: svn://chrome-svn/chromeos/trunk@294 06c00378-0e64-4dae-be16-12b19f9950a1

+++ src/platform/window_manager/window_manager_test.cc
@@ -67,7 +67,7 @@ class WindowManagerTest : public ::testing::Test {
     } else {
       EXPECT_TRUE(exists);
       ASSERT_EQ(expected.size(), actual.size());
-      for (int i = 0; i < actual.size(); ++i)
+      for (size_t i = 0; i < actual.size(); ++i)
         EXPECT_EQ(expected[i], actual[i]);
     }
   }
